+++ mio/core/continuation.py
@@ -3,7 +3,7 @@ from copy import copy
 
 from mio import runtime
 from mio.utils import method, Null
-from mio.types.object import Object
+from mio.object import Object
 
 
 class Continuation(Object):