+++ mio/runtime.py
@@ -3,11 +3,6 @@
 ...
 """
 
-from os import path
-
-
-from pkg_resources import resource_filename, resource_listdir
-
 
 state = None
 
@@ -19,11 +14,6 @@ def init(args=[], opts=None):
 
     state = State(args, opts)
     state.create_objects()
-
-    if opts is None or (opts is not None and not opts.nosys):
-        for resource in resource_listdir(__package__, path.join("lib", "bootstrap")):
-            filename = resource_filename(__package__, path.join("lib", "bootstrap", resource))
-            state.load(filename)
 
 
 def find(name):