+++ mio/types/dict.py
@@ -2,7 +2,7 @@ from copy import copy
 
 
 from mio import runtime
-from mio.types.object import Object
+from mio.object import Object
 from mio.utils import method, Null