+++ mio/types/number.py
@@ -1,7 +1,8 @@
 from decimal import Decimal
 
+
 from mio import runtime
-from mio.types.object import Object
+from mio.object import Object
 from mio.utils import method, Null